refactored with datetime_format classes
