+++ dsgrid/config/date_time_dimension_config.py
@@ -3,13 +3,10 @@ from pyspark.sql.types import (
     StructType,
     StructField,
     TimestampType,
-    DoubleType,
-    StringType,
-    TimestampNTZType,
 )
 import pyspark.sql.functions as F
 
-from dsgrid.dimension.time import make_time_range, TimeZone
+from dsgrid.dimension.time import make_time_range, DatetimeFormat, LeapDayAdjustmentType
 from dsgrid.exceptions import DSGInvalidDataset
 from dsgrid.time.types import DatetimeTimestampType
 from dsgrid.utils.timing import timer_stats_collector, track_timing
@@ -45,46 +42,37 @@ class DateTimeDimensionConfig(TimeDimensionBaseConfig):
         # TODO: need to support validation of multiple time ranges: DSGRID-173
 
         assert (
-            load_data_df.schema[time_col].dataType != TimestampNTZType()
-        ), f"datetime {time_col} column cannot be TimestampNTZType"
+            load_data_df.schema[time_col].dataType == TimestampType()
+        ), f"datetime {time_col} column must be TimestampType"
 
         expected_timestamps = time_range.list_time_range()
-        if self.model.timezone in [TimeZone.LOCAL, TimeZone.LOCAL_MODEL]:
-            if self.model.data_str_format != "":
-                if "id" in load_data_df.columns:
-                    groupby = ["id"]
-                else:
-                    groupby = [
-                        x for x in load_data_df.columns if x not in [time_col, VALUE_COLUMN]
-                    ]
-                start_times = load_data_df.groupBy(*groupby).agg(F.min(time_col).alias(time_col))
-                min_ts = start_times.select(F.min(time_col).alias(time_col)).collect()[0][time_col]
-                start_times = start_times.withColumn("min_ts", F.lit(min_ts))
-                time_deltas = start_times.select(
-                    *groupby, (F.col(time_col) - F.col("min_ts")).alias("time_delta")
-                )
-                timestamps = load_data_df.select(*groupby, time_col).join(
-                    time_deltas, groupby, "left"
-                )
-                timestamps = timestamps.select(
-                    *groupby, (F.col(time_col) - F.col("time_delta")).alias("aligned_time")
-                )
-                timestamp_counts = [
-                    x["count"] for x in timestamps.groupBy("aligned_time").count().collect()
-                ]
-                if len(set(timestamp_counts)) != 1:
-                    raise DSGInvalidDataset(
-                        f"load_data {time_col}s do not have the same time range when adjusting for offsets."
-                    )
-                if len(timestamp_counts) != len(expected_timestamps):
-                    raise DSGInvalidDataset(
-                        f"load_data {time_col}s do not have the expected number of timestamps when adjusting for offsets."
-                    )
-                return
+        if self.model.datetime_format.format_type in [DatetimeFormat.LOCAL_AS_STRINGS]:
+            if "id" in load_data_df.columns:
+                groupby = ["id"]
             else:
-                raise NotImplementedError(
-                    f"Cannot check load_data {time_col}s with Local or LocalModel timezone that relies on geography dimension."
+                groupby = [x for x in load_data_df.columns if x not in [time_col, VALUE_COLUMN]]
+            start_times = load_data_df.groupBy(*groupby).agg(F.min(time_col).alias(time_col))
+            min_ts = start_times.select(F.min(time_col).alias(time_col)).collect()[0][time_col]
+            start_times = start_times.withColumn("min_ts", F.lit(min_ts))
+            time_deltas = start_times.select(
+                *groupby, (F.col(time_col) - F.col("min_ts")).alias("time_delta")
+            )
+            timestamps = load_data_df.select(*groupby, time_col).join(time_deltas, groupby, "left")
+            timestamps = timestamps.select(
+                *groupby, (F.col(time_col) - F.col("time_delta")).alias("aligned_time")
+            )
+            timestamp_counts = [
+                x["count"] for x in timestamps.groupBy("aligned_time").count().collect()
+            ]
+            if len(set(timestamp_counts)) != 1:
+                raise DSGInvalidDataset(
+                    f"load_data {time_col}s do not have the same time range when adjusting for offsets."
                 )
+            if len(timestamp_counts) != len(expected_timestamps):
+                raise DSGInvalidDataset(
+                    f"load_data {time_col}s do not have the expected number of timestamps when adjusting for offsets."
+                )
+            return
 
         actual_timestamps = [
             x[time_col].astimezone().astimezone(tz)
@@ -102,7 +90,7 @@ class DateTimeDimensionConfig(TimeDimensionBaseConfig):
                 f"load_data {time_col}s do not match expected times. mismatch={mismatch}"
             )
 
-    def build_time_dataframe(self, model_years=None, timezone=None, data_adjustment=None):
+    def build_time_dataframe(self, model_years=None, data_adjustment=None):
         # Note: DF.show() displays time in session time, which may be confusing.
         # But timestamps are stored correctly here
 
@@ -110,7 +98,7 @@ class DateTimeDimensionConfig(TimeDimensionBaseConfig):
         assert len(time_col) == 1, time_col
         time_col = time_col[0]
         model_time = self.list_expected_dataset_timestamps(
-            model_years=model_years, timezone=timezone, data_adjustment=data_adjustment
+            model_years=model_years, data_adjustment=data_adjustment
         )
         schema = StructType([StructField(time_col, TimestampType(), False)])
         df_time = get_spark_session().createDataFrame(model_time, schema=schema)
@@ -124,7 +112,7 @@ class DateTimeDimensionConfig(TimeDimensionBaseConfig):
     #     df_time = self.build_time_dataframe()
     #     session_tz = _get_spark_session().conf.get("spark.sql.session.timeZone")
     #     df_time = self._convert_time_zone(
-    #         df_time, time_col, session_tz, self.model.timezone.tz_name
+    #         df_time, time_col, session_tz, self.model.datetime_format.timezone.tz_name
     #     )
 
     #     return df_time
@@ -161,51 +149,11 @@ class DateTimeDimensionConfig(TimeDimensionBaseConfig):
         assert len(ptime_col) == 1, ptime_col
         ptime_col = ptime_col[0]
 
-        if self.model.timezone == TimeZone.LOCAL_MODEL:
-            # if local time zones, create time zone map to covert
-            assert "time_zone" in df.columns, df.columns
-            geo_tz = [
-                TimeZone(row.time_zone) for row in df.select("time_zone").distinct().collect()
-            ]
-            assert geo_tz
-            # for LocalModel time, time represents clock time but is laid out like Standard Time
-            geo_tz2 = [tz.get_standard_time() for tz in geo_tz]
-            schema = StructType(
-                [
-                    StructField("model_time", TimestampType(), False),
-                    StructField(time_col, TimestampType(), False),
-                    StructField("multiplier", DoubleType(), False),
-                    StructField("time_zone", StringType(), False),
-                ]
+        if data_adjustment.leap_day_adjustment != LeapDayAdjustmentType.NONE:
+            df_ptime = project_time_dim.build_time_dataframe(
+                model_years=model_years, data_adjustment=data_adjustment
             )
-            time_map = get_spark_session().createDataFrame([], schema=schema)
-            for tz, tz2 in zip(geo_tz, geo_tz2):
-                # for LocalModel time, table is built in standard time but listed as prevailing
-                index_map = self.build_time_dataframe(
-                    model_years=model_years, timezone=tz2.tz, data_adjustment=data_adjustment
-                )
-                index_map = index_map.withColumn("time_zone", F.lit(tz.value))
-
-                if self.model.timezone == TimeZone.LOCAL_MODEL:
-                    # data_adjustment mapping table
-                    table = self._create_adjustment_map_from_model_time(data_adjustment, tz)
-                    index_map = index_map.selectExpr(
-                        "time_zone", f"{time_col} AS model_time"
-                    ).join(table, ["model_time"], "right")
-                else:
-                    index_map = index_map.withColumn("multiplier", F.lit(1.0))
-                time_map = time_map.union(index_map.select(schema.names))
-
-            nontime_cols = [x for x in df.columns if x != ptime_col]
-            df = (
-                df.selectExpr(*nontime_cols, f"{ptime_col} AS model_time")
-                .join(time_map, on=["model_time", "time_zone"], how="inner")
-                .drop(*["model_time", "time_zone"])
-            )
-            groupby = [x for x in df.columns if x not in [VALUE_COLUMN, "multiplier"]]
-            df = df.groupBy(*groupby).agg(
-                F.sum(F.col(VALUE_COLUMN) * F.col("multiplier")).alias(VALUE_COLUMN)
-            )
+            df = df_ptime.join(df, ptime_col, "inner")
 
         df = self._convert_time_to_project_time_interval(
             df=df, project_time_dim=project_time_dim, wrap_time=wrap_time_allowed
@@ -215,14 +163,10 @@ class DateTimeDimensionConfig(TimeDimensionBaseConfig):
     def get_frequency(self):
         return self.model.frequency
 
-    def get_time_ranges(self, model_years=None, timezone=None, data_adjustment=None):
-        if timezone is None:
-            timezone = self.get_tzinfo()
-        if data_adjustment is None:
-            data_adjustment = DataAdjustmentModel()
+    def get_time_ranges(self, model_years=None, data_adjustment=None):
         ranges = []
         for start, end in self._build_time_ranges(
-            self.model.ranges, self.model.str_format, model_years=model_years, tz=timezone
+            self.model.ranges, self.model.str_format, model_years=model_years, tz=self.get_tzinfo()
         ):
             ranges.append(
                 make_time_range(
@@ -240,30 +184,30 @@ class DateTimeDimensionConfig(TimeDimensionBaseConfig):
         return list(DatetimeTimestampType._fields)
 
     def get_tzinfo(self):
-        return self.model.timezone.tz
+        if self.model.datetime_format.format_type == DatetimeFormat.ALIGNED:
+            return self.model.datetime_format.timezone.tz
+        if self.model.datetime_format.format_type in [DatetimeFormat.LOCAL_AS_STRINGS]:
+            return None
 
     def get_time_interval_type(self):
         return self.model.time_interval_type
 
-    def list_expected_dataset_timestamps(
-        self, model_years=None, timezone=None, data_adjustment=None
-    ):
+    def list_expected_dataset_timestamps(self, model_years=None, data_adjustment=None):
         timestamps = []
         for time_range in self.get_time_ranges(
-            model_years=model_years, timezone=timezone, data_adjustment=data_adjustment
+            model_years=model_years, data_adjustment=data_adjustment
         ):
             timestamps += [DatetimeTimestampType(x) for x in time_range.list_time_range()]
         return timestamps
 
     def _convert_dataset_time_to_datetime(self, load_data_df):
-        if self.model.data_str_format == "":
+        if self.model.datetime_format.format_type != DatetimeFormat.LOCAL_AS_STRINGS:
             return load_data_df
         time_col = self.get_load_data_time_columns()
         assert len(time_col) == 1, time_col
         time_col = time_col[0]
-        nontime_cols = [x for x in load_data_df.columns if x != time_col]
-        load_data_df = load_data_df.select(
-            F.to_timestamp(time_col, self.model.data_str_format).alias(time_col),
-            *nontime_cols,
+        load_data_df = load_data_df.withColumn(
+            time_col,
+            F.to_timestamp(time_col, self.model.datetime_format.data_str_format),
         )
         return load_data_df
