+++ dsgrid/config/noop_time_dimension_config.py
@@ -15,28 +15,34 @@ class NoOpTimeDimensionConfig(TimeDimensionBaseConfig):
     def check_dataset_time_consistency(self, load_data_df, time_columns):
         pass
 
-    def build_time_dataframe(self, model_years=None, timezone=None, data_adjustment=None):
+    def build_time_dataframe(self, model_years=None, data_adjustment=None):
         raise NotImplementedError(f"Cannot build a time dataframe for a {type(self)}")
 
     # def build_time_dataframe_with_time_zone(self):
     #     pass
 
     def convert_dataframe(
-        self, df, project_time_dim, model_years=None, value_columns=None, wrap_time_allowed=False
+        self,
+        df,
+        project_time_dim,
+        model_years=None,
+        value_columns=None,
+        wrap_time_allowed=False,
+        data_adjustment=None,
     ):
         return df
 
     def get_frequency(self):
         return timedelta(days=0)
 
-    def get_time_ranges(self, model_years=None, timezone=None, data_adjustment=None):
+    def get_time_ranges(self, model_years=None, data_adjustment=None):
         frequency = self.get_frequency()
         ranges = [
             make_time_range(
                 start=None,
                 end=None,
                 frequency=frequency,
-                leap_day_adjustment=None,
+                data_adjustment=data_adjustment,
                 time_interval_type=None,
             )
         ]
@@ -52,7 +58,5 @@ class NoOpTimeDimensionConfig(TimeDimensionBaseConfig):
     def get_time_interval_type(self):
         return None
 
-    def list_expected_dataset_timestamps(
-        self, model_years=None, timezone=None, data_adjustment=None
-    ):
+    def list_expected_dataset_timestamps(self, model_years=None, data_adjustment=None):
         return []
