+++ dsgrid/config/representative_period_time_dimension_config.py
@@ -58,13 +58,14 @@ class RepresentativePeriodTimeDimensionConfig(TimeDimensionBaseConfig):
             time_columns,
         )
 
-    def build_time_dataframe(self, model_years=None, timezone=None, data_adjustment=None):
+    def build_time_dataframe(self, model_years=None, data_adjustment=None):
         time_cols = self.get_load_data_time_columns()
         schema = StructType(
             [StructField(time_col, IntegerType(), False) for time_col in time_cols]
         )
-
-        model_time = self.list_expected_dataset_timestamps(model_years=model_years)
+        model_time = self.list_expected_dataset_timestamps(
+            model_years=model_years, data_adjustment=data_adjustment
+        )
         df_time = get_spark_session().createDataFrame(model_time, schema=schema)
 
         return df_time
@@ -73,7 +74,13 @@ class RepresentativePeriodTimeDimensionConfig(TimeDimensionBaseConfig):
     #     return self.build_time_dataframe()
 
     def convert_dataframe(
-        self, df, project_time_dim, model_years=None, value_columns=None, wrap_time_allowed=False
+        self,
+        df,
+        project_time_dim,
+        model_years=None,
+        value_columns=None,
+        wrap_time_allowed=False,
+        data_adjustment=None,
     ):
         if project_time_dim is None:
             return df
@@ -110,7 +117,9 @@ class RepresentativePeriodTimeDimensionConfig(TimeDimensionBaseConfig):
 
         time_df = None
         try:
-            project_time_df = project_time_dim.build_time_dataframe()
+            project_time_df = project_time_dim.build_time_dataframe(
+                model_years=model_years, data_adjustment=data_adjustment
+            )
             map_time = "timestamp_to_map"
             project_time_df = self._shift_time_interval(
                 project_time_df,
@@ -159,9 +168,7 @@ class RepresentativePeriodTimeDimensionConfig(TimeDimensionBaseConfig):
     def get_frequency(self):
         return self._format_handler.get_frequency()
 
-    def get_time_ranges(self, model_years=None, timezone=None, data_adjustment=None):
-        if data_adjustment is None:
-            data_adjustment = DataAdjustmentModel()
+    def get_time_ranges(self, model_years=None, data_adjustment=None):
         if model_years is not None:
             # We do not expect to need this.
             raise NotImplementedError(f"No support for {model_years=} in {type(self)}")
@@ -177,12 +184,12 @@ class RepresentativePeriodTimeDimensionConfig(TimeDimensionBaseConfig):
         return self._format_handler.get_load_data_time_columns()
 
     def get_tzinfo(self):
-        return None  # self.model.timezone.tz
+        return None
 
     def get_time_interval_type(self):
         return self.model.time_interval_type
 
-    def list_expected_dataset_timestamps(self, model_years=None):
+    def list_expected_dataset_timestamps(self, model_years=None, data_adjustment=None):
         if model_years is not None:
             # We do not expect to need this.
             raise NotImplementedError(f"No support for {model_years=} in {type(self)}")
