+++ dsgrid/config/time_dimension_base_config.py
@@ -1,29 +1,15 @@
 import abc
-from datetime import datetime, timedelta
+from datetime import timedelta
 from typing import Optional
-from zoneinfo import ZoneInfo
 
-import pandas as pd
 import pyspark.sql.functions as F
-from pyspark.sql.types import (
-    StructType,
-    StructField,
-    TimestampType,
-    DoubleType,
-)
 
 from .dimension_config import DimensionBaseConfigWithoutFiles
 from dsgrid.dimension.time import TimeZone, TimeIntervalType
-from dsgrid.exceptions import DSGInvalidOperation, DSGInvalidDimension
+from dsgrid.dimension.time_utils import build_time_ranges
+from dsgrid.exceptions import DSGInvalidOperation
 from dsgrid.config.dimensions import TimeRangeModel
-from dsgrid.dimension.time import (
-    DataAdjustmentModel,
-    LeapDayAdjustmentType,
-    DaylightSavingFallBackType,
-    get_dls_fallback_time_change_by_time_range,
-)
-from dsgrid.time.types import DatetimeTimestampType
-from dsgrid.utils.spark import get_spark_session
+from dsgrid.dimension.time import DataAdjustmentModel
 
 
 class TimeDimensionBaseConfig(DimensionBaseConfigWithoutFiles, abc.ABC):
@@ -49,7 +35,6 @@ class TimeDimensionBaseConfig(DimensionBaseConfigWithoutFiles, abc.ABC):
     def build_time_dataframe(
         self,
         model_years: Optional[list[int]] = None,
-        timezone: Optional[ZoneInfo] = None,
         data_adjustment: Optional[DataAdjustmentModel] = None,
     ):
         """Build time dimension as specified in config in a spark dataframe.
@@ -58,6 +43,8 @@ class TimeDimensionBaseConfig(DimensionBaseConfigWithoutFiles, abc.ABC):
         ----------
         model_years : None | list[int]
             If specified, repeat the timestamps for each model year.
+        data_adjustment : None | DataAdjustmentModel
+            Leap day and daylight saving adjustments to make to both time and data.
 
         Returns
         -------
@@ -85,6 +72,7 @@ class TimeDimensionBaseConfig(DimensionBaseConfigWithoutFiles, abc.ABC):
         model_years: Optional[list[int]] = None,
         value_columns: Optional[set[str]] = None,
         wrap_time_allowed: bool = False,
+        data_adjustment: Optional[DataAdjustmentModel] = None,
     ):
         """Convert input df to use project's time format and time zone.
 
@@ -99,6 +87,8 @@ class TimeDimensionBaseConfig(DimensionBaseConfigWithoutFiles, abc.ABC):
             dimension types.
         wrapped_time_allowed : bool
             Whether to allow time-wrapping to align time zone
+        data_adjustment : None | DataAdjustmentModel
+            Leap day and daylight saving adjustments to make to both time and data.
 
         Returns
         -------
@@ -162,7 +152,6 @@ class TimeDimensionBaseConfig(DimensionBaseConfigWithoutFiles, abc.ABC):
     def get_time_ranges(
         self,
         model_years: Optional[list[int]] = None,
-        timezone: Optional[ZoneInfo] = None,
         data_adjustment: Optional[DataAdjustmentModel] = None,
     ):
         """Return time ranges with timezone applied.
@@ -172,8 +161,8 @@ class TimeDimensionBaseConfig(DimensionBaseConfigWithoutFiles, abc.ABC):
         model_years : None | list[int]
             If set, replace the base year in the time ranges with these model years. In this case
             each range must be in the same year.
-        timezone : None | ZoneInfo
-            If set, override the self timezone info when creating the time ranges
+        data_adjustment : None | DataAdjustmentModel
+            Leap day and daylight saving adjustments to make to both time and data.
 
         Returns
         -------
@@ -203,7 +192,11 @@ class TimeDimensionBaseConfig(DimensionBaseConfigWithoutFiles, abc.ABC):
         """
 
     @abc.abstractmethod
-    def list_expected_dataset_timestamps(self, model_years: Optional[list[int]] = None):
+    def list_expected_dataset_timestamps(
+        self,
+        model_years: Optional[list[int]] = None,
+        data_adjustment: Optional[DataAdjustmentModel] = None,
+    ):
         """Return a list of the timestamps expected in the load_data table.
 
         Parameters
@@ -211,6 +204,8 @@ class TimeDimensionBaseConfig(DimensionBaseConfigWithoutFiles, abc.ABC):
         model_years : None | list[int]
             If set, replace the base year in the time ranges with these model years. In this case
             each range must be in the same year.
+        data_adjustment : None | DataAdjustmentModel
+            Leap day and daylight saving adjustments to make to both time and data.
 
         Returns
         -------
@@ -395,129 +390,4 @@ class TimeDimensionBaseConfig(DimensionBaseConfigWithoutFiles, abc.ABC):
         model_years: Optional[list[int]] = None,
         tz: Optional[TimeZone] = None,
     ):
-        ranges = []
-        allowed_year = None
-        for time_range in time_ranges:
-            start = datetime.strptime(time_range.start, str_format)
-            end = datetime.strptime(time_range.end, str_format)
-            if model_years is not None:
-                if start.year != end.year or (
-                    allowed_year is not None and start.year != allowed_year
-                ):
-                    raise DSGInvalidDimension(
-                        f"All time ranges must be in the same year if model_years is set: {model_years=}"
-                    )
-                allowed_year = start.year
-            for year in model_years or [start.year]:
-                start_adj = datetime(
-                    year=year,
-                    month=start.month,
-                    day=start.day,
-                    hour=start.hour,
-                    minute=start.minute,
-                    second=start.second,
-                    microsecond=start.microsecond,
-                )
-                end_adj = datetime(
-                    year=end.year + year - start.year,
-                    month=end.month,
-                    day=end.day,
-                    hour=end.hour,
-                    minute=end.minute,
-                    second=end.second,
-                    microsecond=end.microsecond,
-                )
-                ranges.append((pd.Timestamp(start_adj, tz=tz), pd.Timestamp(end_adj, tz=tz)))
-
-        ranges.sort(key=lambda x: x[0])
-        return ranges
-
-    def _create_adjustment_map_from_model_time(
-        self, data_adjustment: DataAdjustmentModel, time_zone: TimeZone, model_years=None
-    ):
-        """Create data adjustment mapping from model_time to prevailing time (timestamp) of input time_zone."""
-        time_col = list(DatetimeTimestampType._fields)
-        assert len(time_col) == 1, time_col
-        time_col = time_col[0]
-
-        ld_adj = data_adjustment.leap_day_adjustment
-        fb_adj = data_adjustment.daylight_saving_adjustment.fall_back_hour
-
-        TZ_st, TZ_pt = time_zone.get_standard_time(), time_zone.get_prevailing_time()
-        ranges = self.get_time_ranges(
-            model_years=model_years, timezone=TZ_pt.tz, data_adjustment=data_adjustment
-        )
-        freq = self.model.frequency
-        model_time, prevailing_time, multipliers = [], [], []
-        for range in ranges:
-            cur_pt = range.start.to_pydatetime()
-            end_pt = range.end.to_pydatetime()
-
-            if fb_adj == DaylightSavingFallBackType.INTERPOLATE:
-                fb_times = get_dls_fallback_time_change_by_time_range(
-                    cur_pt, end_pt, frequency=freq
-                )  # in PT
-                fb_repeats = [0 for x in fb_times]
-                print(fb_times)
-
-            cur = range.start.to_pydatetime().astimezone(ZoneInfo("UTC"))
-            end = range.end.to_pydatetime().astimezone(ZoneInfo("UTC")) + freq
-
-            while cur < end:
-                multiplier = 1.0
-                frequency = freq
-                cur_pt = cur.astimezone(TZ_pt.tz)
-                model_ts = cur_pt.replace(tzinfo=TZ_st.tz)
-                month = cur_pt.month
-                day = cur_pt.day
-                if ld_adj == LeapDayAdjustmentType.DROP_FEB29 and month == 2 and day == 29:
-                    cur += frequency
-                    pass
-                if ld_adj == LeapDayAdjustmentType.DROP_DEC31 and month == 12 and day == 31:
-                    cur += frequency
-                    pass
-                if ld_adj == LeapDayAdjustmentType.DROP_JAN1 and month == 1 and day == 1:
-                    cur += frequency
-                    pass
-
-                if fb_adj == DaylightSavingFallBackType.INTERPOLATE:
-                    for i, ts in enumerate(fb_times):
-                        if cur == ts.astimezone(ZoneInfo("UTC")):
-                            if fb_repeats[i] == 0:
-                                frequency = timedelta(hours=0)
-                                multiplier = 0.5
-                            if fb_repeats[i] == 1:
-                                model_ts = (
-                                    (cur + timedelta(hours=1))
-                                    .astimezone(TZ_pt.tz)
-                                    .replace(tzinfo=TZ_st.tz)
-                                )
-                                multiplier = 0.5
-                            fb_repeats[i] += 1
-
-                model_time.append(model_ts)
-                prevailing_time.append(cur_pt)
-                multipliers.append(multiplier)
-                cur += frequency
-
-        schema = StructType(
-            [
-                StructField("model_time", TimestampType(), False),
-                StructField(time_col, TimestampType(), False),
-                StructField("multiplier", DoubleType(), False),
-            ]
-        )
-        table = get_spark_session().createDataFrame(
-            zip(model_time, prevailing_time, multipliers), schema=schema
-        )
-        return table
-
-    def _get_tzinfo_from_geography(geography_dim):
-        """Get tzinfo from time_zone column of geography dimension record"""
-        # TODO not currently in use
-        geo_records = geography_dim.get_records_dataframe()
-        geo_tz_values = [
-            row.time_zone for row in geo_records.select("time_zone").distinct().collect()
-        ]
-        geo_tzinfos = [TimeZone(tz).tz for tz in geo_tz_values]
-        return geo_tzinfos
+        return build_time_ranges(time_ranges, str_format, model_years=model_years, tz=tz)
